requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==2.1.0
pydantic==2.6.0
python-dotenv==1.0.0
requests==2.31.0
streamlit==1.22.0
pandas==2.2.3
altair==5.2.1
typing_extensions
